+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Fixed test for `database/sql` extended API 
+* Fixed test for `database/sql` extended API
+* Extend `scheme.Client` interface with method `Database`
 * Fixed `sugar.RecursiveRemove` for remove full path
 * Removed `driver.ResultNoRows` in `internal/xsql`
 * Supported binding parameters for `database/sql` driver by default