+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added integration tests for `database/sql` extended API
 * Fixed test for `database/sql` extended API
 * Extend `scheme.Client` interface with method `Database`
 * Fixed `sugar.RecursiveRemove` for remove full path